Rename type
